QuantumESPRESSOBase

Manual Outline

Installation

QuantumESPRESSOBase.asfieldname

QuantumESPRESSOBase.bravais_lattice

QuantumESPRESSOBase.titleof

QuantumESPRESSOBase.to_qe

QuantumESPRESSOBase.asfieldname — Method

asfieldname(::Type{<:InputEntry})
asfieldname(::InputEntry)

Return the field name of a Namelist or a Card in a QuantumESPRESSOInput.

Examples

julia> asfieldname(SystemNamelist)
:system

julia> asfieldname(SystemNamelist())
:system

QuantumESPRESSOBase.bravais_lattice — Method

bravais_lattice(ibrav::Integer, celldm::AbstractVector{Union{Nothing, Float64}})

Return a 3x3 matrix representing the Bravais lattice from ibrav and celldm.

QuantumESPRESSOBase.titleof — Method

titleof(::Type{<:InputEntry})
titleof(::InputEntry)

Return the title of the input entry in Quantum ESPRESSO.

The definition titleof(x) = titleof(typeof(x)) is provided for convenience so that instances can be passed instead of types.

Examples

julia> titleof(SystemNamelist())
"SYSTEM"

julia> titleof(SystemNamelist)
"SYSTEM"

QuantumESPRESSOBase.to_qe — Method

to_qe(x; indent = ' '^4, delim = ' ')

Return a string representing the object, valid form Quantum ESPRESSO's input.