QuantumESPRESSOBase.Cards module

QuantumESPRESSOBase.Cards — Module

module Cards

Examples

julia>

QuantumESPRESSOBase.Cards.allowed_options — Method

allowed_options(T::Type{<:Card})

Return the allowed options for Card T.

Examples

julia> using QuantumESPRESSOBase.Cards, QuantumESPRESSOBase.Cards.PWscf

julia> allowed_options(AtomicPositionsCard)
("alat", "bohr", "angstrom", "crystal", "crystal_sg")

julia> allowed_options(CellParametersCard)
("alat", "bohr", "angstrom")

julia> allowed_options(KPointsCard)
("tpiba", "automatic", "crystal", "gamma", "tpiba_b", "crystal_b", "tpiba_c", "crystal_c")

QuantumESPRESSOBase.Cards.optionof — Method

optionof(x::Card)

Return the option for Card x.

A user should not use x.option to access a Card's option. Because some Cards do not have an option. Using optionof(x) is suggested.

QuantumESPRESSOBase.Cards.PWscf module

QuantumESPRESSOBase.Cards.PWscf — Module

module PWscf

Examples

julia>

QuantumESPRESSOBase.Cards.PWscf.AndreaDalCorso — Type

AndreaDalCorso <: PseudopotentialFormat

A singleton representing the Andrea Dal Corso's code (old format).

QuantumESPRESSOBase.Cards.PWscf.AtomicPosition — Type

AtomicPosition(atom::Union{AbstractChar,String}, pos::Vector{Float64}[, if_pos::Vector{Int}])
AtomicPosition(atom::Union{AbstractChar,AbstractString})
AtomicPosition(x::AtomicSpecies, pos, if_pos)
AtomicPosition(x::AtomicSpecies)

Represent each line of the ATOMIC_POSITIONS card in QE.

It is a mutable struct and supports incomplete Initialization as in the second and fourth constructors. See the examples below. The atom field accepts at most 3 characters as claimed in QE's documentation.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicPosition('O', [0, 0, 0])
AtomicPosition("O", [0.0, 0.0, 0.0], [1, 1, 1])

julia> x = AtomicPosition('O');

julia> x.pos
ERROR: UndefRefError: access to undefined reference
[...]

julia> x.pos = [0, 0, 0]
ERROR: TypeError: in setfield!, expected Array{Float64,1}, got Array{Int64,1}
[...]

julia> x.pos = Float64[0, 0, 0]
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

julia> x.if_pos = [1, 0, 2]
ERROR: AssertionError: `if_pos` elements must be 0 or 1!
[...]

julia> x.if_pos = [1, 0, 1]
3-element Array{Int64,1}:
 1
 0
 1

julia> x
AtomicPosition("O", [0.0, 0.0, 0.0], [1, 0, 1])

julia> AtomicPosition(
            AtomicSpecies('S', 32.066, "S.pz-n-rrkjus_psl.0.1.UPF"),
            [0.500000000, 0.288675130, 1.974192764],
       )
AtomicPosition("S", [0.5, 0.28867513, 1.974192764], [1, 1, 1])

QuantumESPRESSOBase.Cards.PWscf.AtomicPositionsCard — Type

AtomicPositionsCard{A<:AbstractVector{AtomicPosition}} <: Card

Represent the ATOMIC_POSITIONS card in QE.

Arguments

option::String="alat": allowed values are: "alat", "bohr", "angstrom", "crystal", and "crystal_sg".

data::AbstractVector{AtomicPosition}: A vector containing AtomicPositions.

QuantumESPRESSOBase.Cards.PWscf.AtomicSpecies — Type

AtomicSpecies(atom::Union{AbstractChar,String}, mass::Float64, pseudopot::String)
AtomicSpecies(atom::Union{AbstractChar,AbstractString})
AtomicSpecies(x::AtomicPosition, mass, pseudopot)
AtomicSpecies(x::AtomicPosition)

Represent each line of the ATOMIC_SPECIES card in QE.

It is a mutable struct and supports incomplete Initialization as in the second and fourth constructors. See the examples below. The atom field accepts at most 3 characters as claimed in QE's documentation.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicSpecies("C1", 12, "C.pbe-n-kjpaw_psl.1.0.0.UPF")
AtomicSpecies("C1", 12.0, "C.pbe-n-kjpaw_psl.1.0.0.UPF")

julia> x = AtomicSpecies('S');

julia> x.atom
"S"

julia> x.mass = 32.066; x.mass
32.066

julia> x.pseudopot
ERROR: UndefRefError: access to undefined reference
[...]

julia> x.pseudopot = "S.pz-n-rrkjus_psl.0.1.UPF"; x.pseudopot
"S.pz-n-rrkjus_psl.0.1.UPF"

julia> AtomicSpecies(
           AtomicPosition('S', [0.500000000, 0.288675130, 1.974192764]),
           32.066,
           "S.pz-n-rrkjus_psl.0.1.UPF",
       )
AtomicSpecies("S", 32.066, "S.pz-n-rrkjus_psl.0.1.UPF")

QuantumESPRESSOBase.Cards.PWscf.AtomicSpeciesCard — Type

AtomicSpeciesCard{T<:AbstractVector{<:AtomicSpecies}} <: Card

Represent the ATOMIC_SPECIES card in QE. It does not have an "option".

QuantumESPRESSOBase.Cards.PWscf.OldNormConserving — Type

OldNormConserving <: PseudopotentialFormat

A singleton representing the old PWscf norm-conserving format.

QuantumESPRESSOBase.Cards.PWscf.UnifiedPseudopotentialFormat — Type

UnifiedPseudopotentialFormat <: PseudopotentialFormat

A singleton representing the new UPF format.

If it doesn't work, the pseudopotential format is determined by the file name.

QuantumESPRESSOBase.Cards.PWscf.VanderbiltUltraSoft — Type

VanderbiltUltraSoft <: PseudopotentialFormat

A singleton representing the Vanderbilt US pseudopotential code.

QuantumESPRESSOBase.Cards.PWscf.append_atom! — Method

append_atom!(card::Union{AtomicSpeciesCard,AtomicPositionsCard}, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a AtomicSpeciesCard or AtomicPositionsCard.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.PWscf.append_atom! — Method

append_atom!(v::AbstractVector{AtomicSpecies}, atoms::AbstractVector{<:AbstractString})
append_atom!(v::AbstractVector{AtomicPosition}, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a vector of AtomicSpecies or AtomicPositions.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.PWscf.pseudopot_format — Method

pseudopot_format(data::AtomicSpecies)::String

Return the pseudopotential format.

The pseudopotential file is assumed to be in the new UPF format. If it doesn't work, the pseudopotential format is determined by the file name:

"*.vdb or *.van": Vanderbilt US pseudopotential code

"*.RRKJ3": Andrea Dal Corso's code (old format)

none of the above: old PWscf norm-conserving format

QuantumESPRESSOBase.Cards.PWscf.push_atom! — Method

push_atom!(card::Union{AtomicSpeciesCard,AtomicPositionsCard}, atoms::AbstractString...)

Push an atom or multiple atoms to a AtomicSpeciesCard or AtomicPositionsCard.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.PWscf.push_atom! — Method

push_atom!(v::AbstractVector{AtomicSpecies}, atoms::AbstractString...)
push_atom!(v::AbstractVector{AtomicPosition}, atoms::AbstractString...)

Push an atom or multiple atoms to a vector of AtomicSpecies or AtomicPositions.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.PWscf.PseudopotentialFormat — Type

PseudopotentialFormat

Represent all possible pseudopotential file formats.

QuantumESPRESSOBase.cell_volume — Method

cell_volume(card)

Return the cell volume of a CellParametersCard or RefCellParametersCard, in atomic unit.

QuantumESPRESSOBase.Cards.CP module

QuantumESPRESSOBase.Cards.CP.AndreaDalCorso — Type

AndreaDalCorso <: PseudopotentialFormat

A singleton representing the Andrea Dal Corso's code (old format).

QuantumESPRESSOBase.Cards.CP.AtomicPosition — Type

AtomicPosition(atom::Union{AbstractChar,String}, pos::Vector{Float64}[, if_pos::Vector{Int}])
AtomicPosition(atom::Union{AbstractChar,AbstractString})
AtomicPosition(x::AtomicSpecies, pos, if_pos)
AtomicPosition(x::AtomicSpecies)

Represent each line of the ATOMIC_POSITIONS card in QE.

It is a mutable struct and supports incomplete Initialization as in the second and fourth constructors. See the examples below. The atom field accepts at most 3 characters as claimed in QE's documentation.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicPosition('O', [0, 0, 0])
AtomicPosition("O", [0.0, 0.0, 0.0], [1, 1, 1])

julia> x = AtomicPosition('O');

julia> x.pos
ERROR: UndefRefError: access to undefined reference
[...]

julia> x.pos = [0, 0, 0]
ERROR: TypeError: in setfield!, expected Array{Float64,1}, got Array{Int64,1}
[...]

julia> x.pos = Float64[0, 0, 0]
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

julia> x.if_pos = [1, 0, 2]
ERROR: AssertionError: `if_pos` elements must be 0 or 1!
[...]

julia> x.if_pos = [1, 0, 1]
3-element Array{Int64,1}:
 1
 0
 1

julia> x
AtomicPosition("O", [0.0, 0.0, 0.0], [1, 0, 1])

julia> AtomicPosition(
            AtomicSpecies('S', 32.066, "S.pz-n-rrkjus_psl.0.1.UPF"),
            [0.500000000, 0.288675130, 1.974192764],
       )
AtomicPosition("S", [0.5, 0.28867513, 1.974192764], [1, 1, 1])

QuantumESPRESSOBase.Cards.CP.AtomicPosition — Method

AtomicPosition(x::AtomicVelocity, pos, if_pos)
AtomicPosition(x::AtomicVelocity)

Construct an incomplete AtomicPosition from an AtomicVelocity instance.

QuantumESPRESSOBase.Cards.CP.AtomicPositionsCard — Type

AtomicPositionsCard{A<:AbstractVector{AtomicPosition}} <: Card

Represent the ATOMIC_POSITIONS card in QE.

Arguments

option::String="alat": allowed values are: "alat", "bohr", "angstrom", "crystal", and "crystal_sg".

data::AbstractVector{AtomicPosition}: A vector containing AtomicPositions.

QuantumESPRESSOBase.Cards.CP.AtomicSpecies — Type

AtomicSpecies(atom::Union{AbstractChar,String}, mass::Float64, pseudopot::String)
AtomicSpecies(atom::Union{AbstractChar,AbstractString})
AtomicSpecies(x::AtomicPosition, mass, pseudopot)
AtomicSpecies(x::AtomicPosition)

Represent each line of the ATOMIC_SPECIES card in QE.

It is a mutable struct and supports incomplete Initialization as in the second and fourth constructors. See the examples below. The atom field accepts at most 3 characters as claimed in QE's documentation.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicSpecies("C1", 12, "C.pbe-n-kjpaw_psl.1.0.0.UPF")
AtomicSpecies("C1", 12.0, "C.pbe-n-kjpaw_psl.1.0.0.UPF")

julia> x = AtomicSpecies('S');

julia> x.atom
"S"

julia> x.mass = 32.066; x.mass
32.066

julia> x.pseudopot
ERROR: UndefRefError: access to undefined reference
[...]

julia> x.pseudopot = "S.pz-n-rrkjus_psl.0.1.UPF"; x.pseudopot
"S.pz-n-rrkjus_psl.0.1.UPF"

julia> AtomicSpecies(
           AtomicPosition('S', [0.500000000, 0.288675130, 1.974192764]),
           32.066,
           "S.pz-n-rrkjus_psl.0.1.UPF",
       )
AtomicSpecies("S", 32.066, "S.pz-n-rrkjus_psl.0.1.UPF")

QuantumESPRESSOBase.Cards.CP.AtomicSpecies — Method

AtomicSpecies(x::AtomicVelocity, mass, pseudopot)
AtomicSpecies(x::AtomicVelocity)

Construct an incomplete AtomicSpecies from an AtomicVelocity instance.

QuantumESPRESSOBase.Cards.CP.AtomicSpeciesCard — Type

AtomicSpeciesCard{T<:AbstractVector{<:AtomicSpecies}} <: Card

Represent the ATOMIC_SPECIES card in QE. It does not have an "option".

QuantumESPRESSOBase.Cards.CP.OldNormConserving — Type

OldNormConserving <: PseudopotentialFormat

A singleton representing the old PWscf norm-conserving format.

QuantumESPRESSOBase.Cards.CP.UnifiedPseudopotentialFormat — Type

UnifiedPseudopotentialFormat <: PseudopotentialFormat

A singleton representing the new UPF format.

If it doesn't work, the pseudopotential format is determined by the file name.

QuantumESPRESSOBase.Cards.CP.VanderbiltUltraSoft — Type

VanderbiltUltraSoft <: PseudopotentialFormat

A singleton representing the Vanderbilt US pseudopotential code.

QuantumESPRESSOBase.Cards.CP.append_atom! — Method

append_atom!(v::AbstractVector{AtomicVelocity}, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a vector of AtomicVelocitys.

Note: these new atoms will result in incomplete AtomicVelocitys!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.CP.append_atom! — Method

append_atom!(card::AtomicVelocitiesCard, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a AtomicVelocitiesCard.

Note: these new atoms will result in incomplete AtomicVelocitys!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.CP.append_atom! — Method

append_atom!(card::Union{AtomicSpeciesCard,AtomicPositionsCard}, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a AtomicSpeciesCard or AtomicPositionsCard.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.CP.append_atom! — Method

append_atom!(v::AbstractVector{AtomicSpecies}, atoms::AbstractVector{<:AbstractString})
append_atom!(v::AbstractVector{AtomicPosition}, atoms::AbstractVector{<:AbstractString})

Append a vector of atoms to a vector of AtomicSpecies or AtomicPositions.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: append!, push_atom!

QuantumESPRESSOBase.Cards.CP.pseudopot_format — Method

pseudopot_format(data::AtomicSpecies)::String

Return the pseudopotential format.

The pseudopotential file is assumed to be in the new UPF format. If it doesn't work, the pseudopotential format is determined by the file name:

"*.vdb or *.van": Vanderbilt US pseudopotential code

"*.RRKJ3": Andrea Dal Corso's code (old format)

none of the above: old PWscf norm-conserving format

QuantumESPRESSOBase.Cards.CP.push_atom! — Method

push_atom!(v::AbstractVector{AtomicVelocity}, atoms::AbstractString...)

Push an atom or multiple atoms to a vector of AtomicVelocitys.

Note: these new atoms will result in incomplete AtomicVelocitys!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.CP.push_atom! — Method

push_atom!(card::AtomicVelocitiesCard, atoms::AbstractString...)

Push an atom or multiple atoms to a AtomicVelocitiesCard.

Note: these new atoms will result in incomplete AtomicVelocitys!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.CP.push_atom! — Method

push_atom!(card::Union{AtomicSpeciesCard,AtomicPositionsCard}, atoms::AbstractString...)

Push an atom or multiple atoms to a AtomicSpeciesCard or AtomicPositionsCard.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.CP.push_atom! — Method

push_atom!(v::AbstractVector{AtomicSpecies}, atoms::AbstractString...)
push_atom!(v::AbstractVector{AtomicPosition}, atoms::AbstractString...)

Push an atom or multiple atoms to a vector of AtomicSpecies or AtomicPositions.

Note: these new atoms will result in incomplete AtomicSpecies or AtomicPositions!

See also: push!, append_atom!

QuantumESPRESSOBase.Cards.CP.PseudopotentialFormat — Type

PseudopotentialFormat

Represent all possible pseudopotential file formats.

QuantumESPRESSOBase.cell_volume — Method

cell_volume(card)

Return the cell volume of a CellParametersCard or RefCellParametersCard, in atomic unit.