QuantumESPRESSOBase.Inputs module

QuantumESPRESSOBase.Inputs module

QuantumESPRESSOBase.Inputs.PWscf module

QuantumESPRESSOBase.Inputs.CP module

QuantumESPRESSOBase.Inputs.PHonon module

QuantumESPRESSOBase.Inputs — Module

module Inputs

Examples

julia>

QuantumESPRESSOBase.Inputs.Card — Type

Card <: QuantumESPRESSOInputEntry

The abstraction of all components of a Input that is not a Namelist.

QuantumESPRESSOBase.Inputs.Namelist — Type

Namelist <: QuantumESPRESSOInputEntry

The abstraction of an component of a Input, a basic Fortran data structure.

QuantumESPRESSOBase.Inputs.QuantumESPRESSOInput — Type

Represent input files of executables (such as pw.x and cp.x).

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(nml::Namelist)

Return a String representing a Namelist, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(input::QuantumESPRESSOInput)

Return a String representing a QuantumESPRESSOInput, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.titleof — Method

titleof(::Union{Namelist,Card})

Return the title of the input entry in Quantum ESPRESSO.

The definition titleof(x) = titleof(typeof(x)) is provided for convenience so that instances can be passed instead of types.

Examples

julia> using QuantumESPRESSOBase; using QuantumESPRESSOBase.Inputs.PWscf: ControlNamelist

julia> titleof(ControlNamelist()) == titleof(ControlNamelist) == "CONTROL"
true

QuantumESPRESSOBase.Inputs.allowed_options — Function

allowed_options(T::Type{<:Card})

Return the allowed options for Card T.

Examples

julia> using QuantumESPRESSOBase.Cards, QuantumESPRESSOBase.Cards.PWscf

julia> allowed_options(AtomicPositionsCard)
("alat", "bohr", "angstrom", "crystal", "crystal_sg")

julia> allowed_options(CellParametersCard)
("alat", "bohr", "angstrom")

julia> allowed_options(KPointsCard)
("tpiba", "automatic", "crystal", "gamma", "tpiba_b", "crystal_b", "tpiba_c", "crystal_c")

QuantumESPRESSOBase.Inputs.dropdefault — Method

dropdefault(nml::Namelist)

Return an AbstractDict of non-default values of a Namelist.

QuantumESPRESSOBase.Inputs.getoption — Method

getoption(x::Card)

Return the option for Card x.

Warning

A user should not use x.option to access a Card's option.

QuantumESPRESSOBase.Inputs.PWscf module

QuantumESPRESSOBase.Inputs.PWscf — Module

module PWscf

Examples

julia>

Crystallography.Bravais — Method

Crystallography.Bravais(nml::SystemNamelist)

Return a Bravais from a SystemNamelist.

Crystallography.Lattice — Method

Crystallography.Lattice(nml::SystemNamelist)

Return a Lattice from a SystemNamelist.

QuantumESPRESSOBase.Inputs.PWscf.AbstractCellParametersCard — Type

Represent the abstraction of CELL_PARAMETERS and REF_CELL_PARAMETERS cards in QE.

QuantumESPRESSOBase.Inputs.PWscf.AtomicPosition — Type

AtomicPosition(atom::Union{AbstractChar,String}, pos::Vector{Float64}[, if_pos::Vector{Int}])
AtomicPosition(x::AtomicSpecies, pos, if_pos)

Represent each line of the ATOMIC_POSITIONS card in QE.

The atom field accepts at most 3 characters.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicPosition('O', [0, 0, 0])
AtomicPosition("O", [0.0, 0.0, 0.0], Bool[1, 1, 1])

julia> AtomicPosition(
           AtomicSpecies('S', 32.066, "S.pz-n-rrkjus_psl.0.1.UPF"),
           [0.500000000, 0.288675130, 1.974192764],
       )
AtomicPosition("S", [0.5, 0.28867513, 1.974192764], Bool[1, 1, 1])

QuantumESPRESSOBase.Inputs.PWscf.AtomicPositionsCard — Type

AtomicPositionsCard <: Card

Represent the ATOMIC_POSITIONS card in QE.

Arguments

data::AbstractVector{AtomicPosition}: A vector containing AtomicPositions.

option::String="alat": allowed values are: "alat", "bohr", "angstrom", "crystal", and "crystal_sg".

QuantumESPRESSOBase.Inputs.PWscf.AtomicSpecies — Type

AtomicSpecies(atom::Union{AbstractChar,String}, mass::Float64, pseudopot::String)
AtomicSpecies(x::AtomicPosition, mass, pseudopot)

Represent each line of the ATOMIC_SPECIES card in QE.

The atom field accepts at most 3 characters.

Examples

julia> using QuantumESPRESSOBase.Cards.PWscf

julia> AtomicSpecies("C1", 12, "C.pbe-n-kjpaw_psl.1.0.0.UPF")
AtomicSpecies("C1", 12.0, "C.pbe-n-kjpaw_psl.1.0.0.UPF")

julia> AtomicSpecies(
           AtomicPosition('S', [0.500000000, 0.288675130, 1.974192764]),
           32.066,
           "S.pz-n-rrkjus_psl.0.1.UPF",
       )
AtomicSpecies("S", 32.066, "S.pz-n-rrkjus_psl.0.1.UPF")

QuantumESPRESSOBase.Inputs.PWscf.AtomicSpeciesCard — Type

AtomicSpeciesCard <: Card

Represent the ATOMIC_SPECIES card in QE. It does not have an "option".

QuantumESPRESSOBase.Inputs.PWscf.BandsNamelist — Type

BandsNamelist <: Namelist

Represent the BANDS namelist of bands.x.

QuantumESPRESSOBase.Inputs.PWscf.CellNamelist — Type

CellNamelist <: Namelist
CellNamelist(cell_dynamics, press, wmass, cell_factor, press_conv_thr, cell_dofree)
CellNamelist(; kwargs...)
CellNamelist(::CellNamelist, kwargs...)
CellNamelist(::CellNamelist, dict)

Represent the CELL namelist of pw.x.

Input this namelist only if calculation is "vc-relax" or "vc-md".

QuantumESPRESSOBase.Inputs.PWscf.CellParametersCard — Type

CellParametersCard{T<:Real} <: AbstractCellParametersCard
CellParametersCard(data::AbstractMatrix, option::String)

Represent the CELL_PARAMETERS cards in PWscf and CP packages.

QuantumESPRESSOBase.Inputs.PWscf.ControlNamelist — Type

ControlNamelist <: Namelist
ControlNamelist(calculation, title, verbosity, restart_mode, wf_collect, nstep, iprint, tstress, tprnfor, dt, outdir, wfcdir, prefix, lkpoint_dir, max_seconds, etot_conv_thr, forc_conv_thr, disk_io, pseudo_dir, tefield, dipfield, lelfield, nberrycyc, lorbm, lberry, gdir, nppstr, lfcpopt, gate)
ControlNamelist(; kwargs...)
ControlNamelist(::ControlNamelist, kwargs...)
ControlNamelist(::ControlNamelist, dict)

Represent the CONTROL namelist of pw.x.

QuantumESPRESSOBase.Inputs.PWscf.DosNamelist — Type

DosNamelist <: Namelist

Represent the DOS namelist of dos.x.

QuantumESPRESSOBase.Inputs.PWscf.ElectronsNamelist — Type

ElectronsNamelist <: Namelist
ElectronsNamelist(electron_maxstep, scf_must_converge, conv_thr, adaptive_thr, conv_thr_init, conv_thr_multi, mixing_mode, mixing_beta, mixing_ndim, mixing_fixed_ns, diagonalization, ortho_para, diago_thr_init, diago_cg_maxiter, diago_david_ndim, diago_full_acc, efield, efield_cart, efield_phase, startingpot, startingwfc, tqr)
ElectronsNamelist(; kwargs...)
ElectronsNamelist(::ElectronsNamelist, kwargs...)
ElectronsNamelist(::ElectronsNamelist, dict)

Represent the ELECTRONS namelist of pw.x.

QuantumESPRESSOBase.Inputs.PWscf.GammaPoint — Type

Represent the centre of the Brillouin zone (commonly marked as the Γ point).

QuantumESPRESSOBase.Inputs.PWscf.IonsNamelist — Type

IonsNamelist <: Namelist
IonsNamelist(ion_dynamics, ion_positions, pot_extrapolation, wfc_extrapolation, remove_rigid_rot, ion_temperature, tempw, tolp, delta_t, nraise, refold_pos, upscale, bfgs_ndim, trust_radius_max, trust_radius_min, trust_radius_ini, w_1, w_2)
IonsNamelist(; kwargs...)
IonsNamelist(::IonsNamelist, kwargs...)
IonsNamelist(::IonsNamelist, dict)

Represent the IONS namelist of pw.x.

Input this namelist only if calculation is "relax", "md", "vc-relax", or "vc-md".

QuantumESPRESSOBase.Inputs.PWscf.KPointsCard — Type

struct KPointsCard{<:Union{MonkhorstPackGrid,GammaPoint,AbstractVector{SpecialKPoint}}} <: Card

Represent the K_POINTS card in QE.

Arguments

data::Union{MonkhorstPackGrid,GammaPoint,AbstractVector{SpecialKPoint}}: A Γ point, a Monkhorst–Pack grid or a vector containing SpecialKPoints.

option::String="tpiba": allowed values are: "tpiba", "automatic", "crystal", "gamma", "tpibab", "crystalb", "tpibac" and "crystalc".

QuantumESPRESSOBase.Inputs.PWscf.MonkhorstPackGrid — Type

MonkhorstPackGrid(mesh, is_shift)

Represent the Monkhorst–Pack grid.

Arguments

mesh: A length-three vector specifying the k-point grid ($nk_1 × nk_2 × nk_3$) as in Monkhorst–Pack grids.

is_shift: A length-three vector specifying whether the grid is displaced by half a grid step in the corresponding directions.

QuantumESPRESSOBase.Inputs.PWscf.PWInput — Type

PWInput <: QuantumESPRESSOInput
PWInput(control, system, electrons, ions, cell, atomic_species, atomic_positions, k_points, cell_parameters)

Construct a PWInput which represents the input of program pw.x.

Arguments

control::ControlNamelist=ControlNamelist(): the CONTROL namelist of the input. Optional.

system::SystemNamelist=SystemNamelist(): the SYSTEM namelist of the input. Optional.

electrons::ElectronsNamelist=ElectronsNamelist(): the ELECTRONS namelist of the input. Optional.

ions::IonsNamelist=IonsNamelist(): the IONS namelist of the input. Optional.

cell::CellNamelist=CellNamelist(): the CELL namelist of the input. Optional.

atomic_species::AtomicSpeciesCard: the ATOMIC_SPECIES card of the input. Must be provided explicitly.

atomic_positions::AtomicPositionsCard: the ATOMIC_POSITIONS card of the input. Must be provided explicitly.

k_points::KPointsCard: the K_POINTS card of the input. Must be provided explicitly.

cell_parameters::Union{Nothing,CellParametersCard}: the CELL_PARAMETERS card of the input. Must be either nothing or a CellParametersCard.

QuantumESPRESSOBase.Inputs.PWscf.SpecialKPoint — Type

SpecialKPoint(coord, weight)

Represent a special point of the 3D Brillouin zone. Each of them has a weight.

QuantumESPRESSOBase.Inputs.PWscf.SystemNamelist — Type

SystemNamelist <: Namelist
SystemNamelist(ibrav, celldm, A, B, C, cosAB, cosAC, cosBC, nat, ntyp, nbnd, tot_charge, starting_charge, tot_magnetization, starting_magnetization, ecutwfc, ecutrho, ecutfock, nr1, nr2, nr3, nr1s, nr2s, nr3s, nosym, nosym_evc, noinv, no_t_rev, force_symmorphic, use_all_frac, occupations, one_atom_occupations, starting_spin_angle, degauss, smearing, nspin, noncolin, ecfixed, qcutz, q2sigma, input_dft, exx_fraction, screening_parameter, exxdiv_treatment, x_gamma_extrapolation, ecutvcut, nqx1, nqx2, nqx3, localization_thr, lda_plus_u, lda_plus_u_kind, Hubbard_U, Hubbard_J0, Hubbard_alpha, Hubbard_beta, starting_ns_eigenvalue, U_projection_type, edir, emaxpos, eopreg, eamp, angle1, angle2, constrained_magnetization, fixed_magnetization, lambda, report, lspinorb, assume_isolated, esm_bc, esm_w, esm_efield, esm_nfit, fcp_mu, vdw_corr, london, london_s6, london_c6, london_rvdw, london_rcut, ts_vdw_econv_thr, ts_vdw_isolated, xdm, xdm_a1, xdm_a2, space_group, uniqueb, origin_choice, rhombohedral, zgate, relaxz, block, block_1, block_2, block_height)
SystemNamelist(; kwargs...)
SystemNamelist(::SystemNamelist, kwargs...)
SystemNamelist(::SystemNamelist, dict)

Represent the SYSTEM namelist of pw.x.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(card::AtomicPositionsCard)

Return a String representing a AtomicPositionsCard, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(data::AtomicPosition)

Return a String representing a AtomicPosition, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(card::AtomicSpeciesCard)

Return a String representing a AtomicSpeciesCard, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(data::AtomicSpecies)

Return a String representing a AtomicSpecies, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(card::CellParametersCard)

Return a String representing a CellParametersCard, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(data::GammaPoint)

Return a String representing a GammaPoint, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(card::KPointsCard)

Return a String representing a KPointsCard, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(data::MonkhorstPackGrid)

Return a String representing a MonkhorstPackGrid, valid for Quantum ESPRESSO's input.

AbInitioSoftwareBase.Inputs.inputstring — Method

inputstring(data::SpecialKPoint)

Return a String representing a SpecialKPoint, valid for Quantum ESPRESSO's input.

Crystallography.cellvolume — Method

cellvolume(card)

Return the cell volume of a CellParametersCard or RefCellParametersCard, in atomic unit.

Warning

It will throw an error if the option is "alat".

Crystallography.cellvolume — Method

cellvolume(input::PWInput)

Return the volume of the cell based on the information given in a PWInput, in atomic unit.

Crystallography.cellvolume — Method

cellvolume(nml::SystemNamelist)

Return the volume of the cell based on the information given in a SystemNamelist, in atomic unit.

Pseudopotentials.pseudopot_format — Method

pseudopot_format(data::AtomicSpecies)

Return the pseudopotential format of the AtomicSpecies.

QuantumESPRESSOBase.Inputs.PWscf.optconvert — Method

optconvert(new_option::AbstractString, card::AbstractCellParametersCard)

Convert the option of an AbstractCellParametersCard from "bohr" to "angstrom", or its reverse.

Warning

It does not support conversion between "alat" and the others.

QuantumESPRESSOBase.Inputs.PWscf.set_temperature — Method

set_temperature(system::PWInput, temperature)

Return a modified PWInput, with finite temperature set.

Warning

Can be used with(out) units. If no unit is given, "Ry" is chosen.

QuantumESPRESSOBase.Inputs.PWscf.set_temperature — Method

set_temperature(system::SystemNamelist, temperature)

Return a modified SystemNamelist, with finite temperature set.

Warning

Can be used with(out) units. If no unit is given, "Ry" is chosen.

QuantumESPRESSOBase.Inputs.PWscf.set_verbosity — Method

set_verbosity(template::ControlNamelist, verbosity)

Return a modified ControlNamelist, with verbosity set.

QuantumESPRESSOBase.Inputs.PWscf.set_verbosity — Method

set_verbosity(template::PWInput, verbosity)

Return a modified PWInput, with verbosity set.

QuantumESPRESSOBase.Inputs.CP module

Modules = [QuantumESPRESSOBase.Inputs.CP]

QuantumESPRESSOBase.Inputs.PHonon module

QuantumESPRESSOBase.Inputs.PHonon — Module

module PHonon

Examples

julia>

QuantumESPRESSOBase.Inputs.PHonon.DynmatNamelist — Type

DynmatNamelist <: Namelist

Represent the INPUT namelist of dynmat.x.

QuantumESPRESSOBase.Inputs.PHonon.MatdynNamelist — Type

MatdynNamelist <: Namelist

Represent the INPUT namelist of matdyn.x.

QuantumESPRESSOBase.Inputs.PHonon.PhNamelist — Type

PhNamelist <: Namelist

Represent the INPUTPH namelist of ph.x.

QuantumESPRESSOBase.Inputs.PHonon.Q2rNamelist — Type

Q2rNamelist <: Namelist

Represent the INPUT namelist of q2r.x.