QuantumESPRESSOBase.Inputs module

QuantumESPRESSOBase.Inputs module

Types

Methods

Types

Methods

QuantumESPRESSOBase.Inputs.getoption — Function

getoption(card::Card)

Return a String representing the option of a Card.

Warning

Do not use card.option to access the option since it may not exist.

QuantumESPRESSOBase.Inputs.optionpool — Function

optionpool(card::Card)
optionpool(T::Type{<:Card})

Return the allowed options for a Card or a Card type.

Examples

julia> optionpool(AtomicPositionsCard)
("alat", "bohr", "angstrom", "crystal", "crystal_sg")

julia> optionpool(CellParametersCard)
("alat", "bohr", "angstrom")

julia> optionpool(SpecialKPointsCard)
("tpiba", "crystal", "tpiba_b", "crystal_b", "tpiba_c", "crystal_c")

Missing docstring.

Missing docstring for allnamelists. Check Documenter's build log for details.

Missing docstring.

Missing docstring for allcards. Check Documenter's build log for details.

Missing docstring.

Missing docstring for required_namelists. Check Documenter's build log for details.

Missing docstring.

Missing docstring for optional_namelists. Check Documenter's build log for details.

Missing docstring.

Missing docstring for required_cards. Check Documenter's build log for details.

Missing docstring.

Missing docstring for optional_cards. Check Documenter's build log for details.