QuantumESPRESSOBase.Inputs.PHonon module

QuantumESPRESSOBase.Inputs.PHonon module

Types

Methods

Types

QuantumESPRESSOBase.Inputs.PHonon.PhNamelist — Type

PhNamelist(amass, outdir, prefix, niter_ph, tr2_ph, alpha_mix, nmix_ph, verbosity, reduce_io, max_seconds, fildyn, fildrho, fildvscf, epsil, lrpa, lnoloc, trans, lraman, eth_rps, eth_ns, dek, recover, low_directory_check, only_init, qplot, q2d, q_in_band_form, electron_phonon, lshift_q, zeu, zue, elop, fpol, ldisp, nogg, asr, ldiag, lqdir, search_sym, nq1, nq2, nq3, nk1, nk2, nk3, k1, k2, k3, start_irr, last_irr, nat_todo, modenum, start_q, last_q, dvscf_star, drho_star)
PhNamelist(; kwargs...)
PhNamelist(::PhNamelist; kwargs...)

Represent the INPUTPH namelist of ph.x.

QuantumESPRESSOBase.Inputs.PHonon.Q2rNamelist — Type

Q2rNamelist(fildyn, flfrc, loto_2d, zasr)
Q2rNamelist(; kwargs...)
Q2rNamelist(::Q2rNamelist; kwargs...)

Represent the INPUT namelist of q2r.x.

QuantumESPRESSOBase.Inputs.PHonon.MatdynNamelist — Type

MatdynNamelist(dos, deltaE, ndos, nk1, nk2, nk3, asr, readtau, flfrc, fldos, flfrq, flvec, fleig, fldyn, fltau, amass, at, ntyp, l1, l2, l3, la2F, q_in_band_form, eigen_similarity, q_in_cryst_coord, na_ifc, fd, nosym, loto_2d)
MatdynNamelist(; kwargs...)
MatdynNamelist(::MatdynNamelist; kwargs...)

Represent the INPUT namelist of matdyn.x.

QuantumESPRESSOBase.Inputs.PHonon.DynmatNamelist — Type

DynmatNamelist(asr, axis, fildyn, filout, filmol, filxsf, fileig, amass, q, lperm, lplasma)
DynmatNamelist(; kwargs...)
DynmatNamelist(::DynmatNamelist; kwargs...)

Represent the INPUT namelist of dynmat.x.

Methods

QuantumESPRESSOBase.Inputs.PHonon.relayinfo — Function

relayinfo(from::PWInput, to::PhInput)

Relay shared information from a PWInput to a PhInput.

A PWInput before a PhInput has the information of outdir and prefix. They must keep the same in a phonon calculation.

relayinfo(from::PhInput, to::Q2rInput)

Relay shared information from a PhInput to a Q2rInput.

A PhInput before a Q2rInput has the information of fildyn. It must keep the same in a q2r calculation.

relayinfo(from::Q2rInput, to::MatdynInput)

Relay shared information from a Q2rInput to a MatdynInput.

A Q2rInput before a MatdynInput has the information of fildyn, flfrc and loto_2d. They must keep the same in a matdyn calculation.

relayinfo(from::PhInput, to::DynmatInput)

Relay shared information from a PhInput to a DynmatInput.

A PhInput before a DynmatInput has the information of asr, fildyn and amass. They must keep the same in a dynmat calculation.