API

QuantumESPRESSOBase module

Crystallography.Lattice — Method

Crystallography.Lattice(::Bravais, p[, obverse::Bool])

Create a Bravais lattice from the exact lattice type and cell parameters p (not celldm!).

The first elements of p are a, b, c; the last 3 are α, β, γ (in radians).