QuantumESPRESSOBase.Inputs module

QuantumESPRESSOBase.Inputs — Module

module Inputs

Examples

julia>

QuantumESPRESSOBase.Inputs.cards — Method

cards(input::QuantumESPRESSOInput)

Return an iterable of Cards of a QuantumESPRESSOInput. It is lazy, you may want to collect it.

QuantumESPRESSOBase.Inputs.compulsory_cards — Method

compulsory_cards(input::CPInput)

Return an iterable of compulsory Cards of a CPInput (AtomicSpeciesCard and AtomicPositionsCard). It is lazy, you may want to collect it.

To get the optional Cards, use (!compulsory_cards)(input).

QuantumESPRESSOBase.Inputs.compulsory_cards — Method

compulsory_cards(input::PWInput)

Return an iterable of compulsory Cards of a PWInput (AtomicSpeciesCard, AtomicPositionsCard and KPointsCard). It is lazy, you may want to collect it.

To get the optional Cards, use (!compulsory_cards)(input).

QuantumESPRESSOBase.Inputs.compulsory_namelists — Method

compulsory_namelists(input::Union{PWInput,CPInput})

Return an iterable of compulsory Namelists of a PWInput or CPInput (ControlNamelist, SystemNamelist and ElectronsNamelist). It is lazy, you may want to collect it.

To get the optional Namelists, use (!compulsory_namelists)(input).

QuantumESPRESSOBase.Inputs.namelists — Method

namelists(input::QuantumESPRESSOInput)

Return an iterable of Namelists of a QuantumESPRESSOInput. It is lazy, you may want to collect it.

QuantumESPRESSOBase.Inputs.PWscf module

QuantumESPRESSOBase.Inputs.PWscf — Module

module PWscf

Examples

julia>

QuantumESPRESSOBase.Inputs.PWscf.PWInput — Type

PWInput(control, system, electrons, ions, cell, atomic_species, atomic_positions, k_points, cell_parameters)

Construct a PWInput which represents the input of program pw.x.

Arguments

control::ControlNamelist=ControlNamelist(): the CONTROL namelist of the input. Optional.

system::SystemNamelist=SystemNamelist(): the SYSTEM namelist of the input. Optional.

electrons::ElectronsNamelist=ElectronsNamelist(): the ELECTRONS namelist of the input. Optional.

ions::IonsNamelist=IonsNamelist(): the IONS namelist of the input. Optional.

cell::CellNamelist=CellNamelist(): the CELL namelist of the input. Optional.

atomic_species::AtomicSpeciesCard: the ATOMIC_SPECIES card of the input. Must be provided explicitly.

atomic_positions::AtomicPositionsCard: the ATOMIC_POSITIONS card of the input. Must be provided explicitly.

k_points::KPointsCard: the K_POINTS card of the input. Must be provided explicitly.

cell_parameters::Union{Nothing,CellParametersCard}: the CELL_PARAMETERS card of the input. Must be either nothing or a CellParametersCard.

QuantumESPRESSOBase.Inputs.CP module

QuantumESPRESSOBase.Inputs.PHonon module

QuantumESPRESSOBase.Inputs.PHonon — Module

module PHonon

Examples

julia>