API

QuantumESPRESSOBase module

QuantumESPRESSOBase.asfieldname — Method

asfieldname(::Type{<:InputEntry})
asfieldname(::InputEntry)

Return the field name of a Namelist or a Card in a QuantumESPRESSOInput.

Examples

julia> using QuantumESPRESSOBase; using QuantumESPRESSOBase.Namelists.PWscf: SystemNamelist

julia> asfieldname(SystemNamelist) == asfieldname(SystemNamelist()) == :system
true

QuantumESPRESSOBase.bravais_lattice — Method

bravais_lattice(ibrav::Integer, celldm::AbstractVector{Union{Nothing, Float64}})

Return a 3x3 matrix representing the Bravais lattice from ibrav and celldm.

QuantumESPRESSOBase.titleof — Method

titleof(::Type{<:InputEntry})
titleof(::InputEntry)

Return the title of the input entry in Quantum ESPRESSO.

The definition titleof(x) = titleof(typeof(x)) is provided for convenience so that instances can be passed instead of types.

Examples

julia> using QuantumESPRESSOBase; using QuantumESPRESSOBase.Namelists.PWscf: SystemNamelist

julia> titleof(SystemNamelist()) == titleof(SystemNamelist) == "SYSTEM"
true

QuantumESPRESSOBase.to_qe — Method

to_qe(x; indent = ' '^4, delim = ' ')

Return a string representing the object, valid form Quantum ESPRESSO's input.